fix(mobile): rename native app to DevCard
