+++ apps/mobile/android/app/src/main/java/com/devcard/app/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.rntest
+package com.devcard.app
 
 import com.facebook.react.ReactActivity
 import com.facebook.react.ReactActivityDelegate
@@ -11,7 +11,7 @@ class MainActivity : ReactActivity() {
    * Returns the name of the main component registered from JavaScript. This is used to schedule
    * rendering of the component.
    */
-  override fun getMainComponentName(): String = "RNTest"
+  override fun getMainComponentName(): String = "DevCard"
 
   /**
    * Returns the instance of the [ReactActivityDelegate]. We use [DefaultReactActivityDelegate]
