+++ apps/mobile/android/app/src/main/java/com/devcard/app/MainApplication.kt
@@ -1,4 +1,4 @@
-package com.rntest
+package com.devcard.app
 
 import android.app.Application
 import com.facebook.react.PackageList
