+++ apps/mobile/app.json
@@ -1,5 +1,5 @@
 {
-  "name": "RNTest",
+  "name": "DevCard",
   "displayName": "DevCard",
   "scheme": "devcard"
 }
