+++ apps/mobile/ios/DevCard.xcodeproj/xcshareddata/xcschemes/DevCard.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "RNTest.app"
-               BlueprintName = "RNTest"
-               ReferencedContainer = "container:RNTest.xcodeproj">
+               BuildableName = "DevCard.app"
+               BlueprintName = "DevCard"
+               ReferencedContainer = "container:DevCard.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "RNTestTests.xctest"
-               BlueprintName = "RNTestTests"
-               ReferencedContainer = "container:RNTest.xcodeproj">
+               BuildableName = "DevCardTests.xctest"
+               BlueprintName = "DevCardTests"
+               ReferencedContainer = "container:DevCard.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "RNTest.app"
-            BlueprintName = "RNTest"
-            ReferencedContainer = "container:RNTest.xcodeproj">
+            BuildableName = "DevCard.app"
+            BlueprintName = "DevCard"
+            ReferencedContainer = "container:DevCard.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "RNTest.app"
-            BlueprintName = "RNTest"
-            ReferencedContainer = "container:RNTest.xcodeproj">
+            BuildableName = "DevCard.app"
+            BlueprintName = "DevCard"
+            ReferencedContainer = "container:DevCard.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
