+++ apps/mobile/ios/DevCard/AppDelegate.swift
@@ -24,7 +24,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     window = UIWindow(frame: UIScreen.main.bounds)
 
     factory.startReactNative(
-      withModuleName: "RNTest",
+      withModuleName: "DevCard",
       in: window,
       launchOptions: launchOptions
     )
