+++ apps/mobile/ios/DevCard/Images.xcassets/AppIcon.appiconset/Contents.json
