+++ apps/mobile/ios/DevCard/Images.xcassets/Contents.json
