+++ apps/mobile/ios/DevCard/LaunchScreen.storyboard
