+++ apps/mobile/ios/DevCard/PrivacyInfo.xcprivacy
