+++ apps/mobile/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@devcard/mobile-new",
+  "name": "@devcard/mobile",
   "version": "0.0.1",
   "private": true,
   "scripts": {
